spiderMonitor-v1.5/requirements.txt
Acquisition==4.4.4
alabaster==0.7.10
alipay==0.7.4
aniso8601==1.2.0
app-version==1.0.1
APScheduler==3.3.1
argh==0.26.2
asn1crypto==0.24.0
attrs==16.3.0
AuthEncoding==4.0.0
Automat==0.6.0
Babel==2.5.1
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
bitarray==0.8.1
blinker==1.4
BTrees==4.4.1
bumpversion==0.5.3
captcha==0.2.4
certifi==2017.7.27.1
cffi==1.9.1
Chameleon==3.2
characteristic==14.3.0
chardet==3.0.4
check-manifest==0.36
chromedriver==2.24.1
click==6.7
colorama==0.3.9
ConcurrentLogHandler==0.9.1
configparser==3.5.0
constantly==15.1.0
coverage==4.3.4
crypto==1.4.1
cryptography==1.8.1
cssselect==1.0.1
DateTime==4.2
defusedxml==0.5.0
Django==2.0.1
django-cors-headers==2.1.0
docker==2.4.2
docker-pycreds==0.2.1
docopt==0.6.2
docutils==0.14
docx==0.2.4
elasticsearch==2.3.0
enum-compat==0.0.2
enum34==1.1.6
ExtensionClass==4.3.0
first==2.0.1
Flask==0.12
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.3
Flask-JWT-Extended==3.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.5
flask-restful-swagger==0.19
Flask-SQLAlchemy==2.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
gerapy==0.7.7
gi==1.2
gunicorn==19.7.1
hyperlink==17.3.1
idna==2.5
imagesize==0.7.1
incremental==17.5.0
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.5
jsonschema==2.6.0
kafka==1.3.5
kafka-python==1.3.3
kazoo==2.2.1
lru-cache==0.2.3
lxml==3.7.3
MarkupSafe==1.0
mock==2.0.0
MultiMapping==4.0
mysql-connector-python==8.0.6
MySQL-python==1.2.5
Naked==0.1.31
nose==1.3.7
notify==0.3.0
numpy==1.13.3
olefile==0.44
packaging==16.8
parsel==1.1.0
pathtools==0.1.2
pbr==2.0.0
pdfkit==0.6.1
Persistence==2.13.2
persistent==4.2.4.2
pika==0.11.0
PIL==1.1.7
Pillow==4.2.1
pip-tools==1.11.0
pkginfo==1.4.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pybloom==1.1
pycparser==2.17
pycrypto==2.6
pycurl==7.43.0
PyDispatcher==2.0.5
Pygments==2.2.0
PyJWT==1.6.0
pykafka==2.7.0
pymango==0.1.1
pymongo==3.5.0
PyMySQL==0.7.11
pyOpenSSL==16.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
pypiwin32==219
pyquery==1.2.17
pyspider==0.3.9
python-dateutil==2.6.0
python-docx==0.8.6
python-json-logger==0.1.7
python-redis-lock==3.2.0
python-scrapyd-api==2.0.1
pytz==2017.2
pywin32==221
PyYAML==3.12
queuelib==1.4.2
redis==2.10.5
reportlab==3.4.0
requests==2.13.0
requests-file==1.4.1
requests-toolbelt==0.8.0
RestrictedPython==3.6.0
retrying==1.3.3
Scrapy==1.3.3
scrapy-redis==0.7.0.dev0
scrapy-splash==0.7.2
scrapyd==1.2.0
scrapyd-client==1.1.0
scutils==1.2.0
selenium==3.8.0
service-identity==16.0.0
shellescape==3.4.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
SpiderKeeper==1.2.0
SQLAlchemy==1.1.9
tabulate==0.8.2
tblib==1.3.2
testfixtures==4.13.5
tldextract==2.0.2
tornado==4.5.1
tqdm==4.19.5
transaction==2.1.2
trollius==2.1
twine==1.9.1
Twisted==17.1.0
typing==3.6.2
tzlocal==1.3
u-msgpack-python==2.4.1
ujson==1.35
urllib3==1.22
validate-email==1.3
vboxapi==1.0
virtualenv==15.1.0
w3lib==1.17.0
waitress==1.1.0
watchdog==0.8.3
WebOb==1.7.4
websocket-client==0.44.0
WebTest==2.0.29
Werkzeug==0.12.1
wkhtmltopdf==0.2
WsgiDAV==2.2.4
WSGIProxy2==0.4.4
z3c.pt==3.1.0
zc.lockfile==1.2.1
ZConfig==3.2.0
zdaemon==4.2.0
ZEO==5.1.0
zExceptions==3.6.1
ZODB==5.3.0
ZODB3==3.11.0
zodbpickle==0.7.0
zope.browser==2.2.0
zope.browserpage==4.2.0
zope.browserresource==4.2.1
zope.cachedescriptors==4.3.1
zope.component==4.4.1
zope.configuration==4.1.0
zope.container==4.2.1
zope.contentprovider==4.1.0
zope.contenttype==4.3.0
zope.deferredimport==4.2.1
zope.dottedname==4.2
zope.event==4.3.0
zope.exceptions==4.2.0
zope.filerepresentation==4.2.0
zope.i18n==4.2.0
zope.i18nmessageid==4.1.0
zope.interface==4.3.3
zope.lifecycleevent==4.2.0
zope.location==4.1.0
zope.pagetemplate==4.3.0
zope.processlifetime==2.2.0
zope.proxy==4.3.0
zope.publisher==4.3.2
zope.schema==4.5.0
zope.security==4.2.1
zope.size==4.2.0
zope.tal==4.3.0
zope.tales==4.2.0
zope.testbrowser==5.2.4
zope.testing==4.6.2
zope.traversing==4.2.0
zope.viewlet==4.1.0
